+++ eslint.config.mjs
@@ -34,6 +34,15 @@ export default defineConfig([
                     selector: "ImportDeclaration[source.value='@testing-library/react']",
                     message: "Please use wrapper from 'test-utils/utils.tsx'.",
                 },
+                {
+                    selector: "MemberExpression[object.name='React'][property.name='useId']",
+                    message: "Please use 'src/hooks/useUniqId' instead.",
+                },
+                {
+                    selector:
+                        "MemberExpression[object.name='React'][property.name='useLayoutEffect']",
+                    message: "Please use 'src/hooks/useLayoutEffect' instead.",
+                },
             ],
             'jsx-a11y/no-autofocus': 'off',
             'import/no-extraneous-dependencies': 'off',