+++ src/components/Select/types.ts
@@ -90,7 +90,7 @@ export type SelectSize = InputControlSize;
 export type SelectRenderCounter = (
     originalComponent: React.ReactElement<SelectCounterProps>,
     counterProps: SelectCounterProps,
-) => React.ReactElement;
+) => React.ReactNode;
 
 export type SelectProps<T = any> = AriaLabelingProps &
     QAProps &
@@ -164,8 +164,10 @@ export type SelectOption<T = any> = QAProps &
     };
 
 export type SelectOptionGroup<T = any> = {
-    /** Label is a string which displayed above the options group.
-     * If label is empty string, group item height will be 0 and only border will be displayed */
+    /**
+     * Label is a string which displayed above the options group.
+     * If label is empty string, group item height will be 0 and only border will be displayed
+     */
     label: string;
     data?: T;
     options?: SelectOption<T>[];