+++ src/components/mobile/index.ts
@@ -2,5 +2,6 @@ export * from './MobileContext';
 export * from './MobileProvider';
 export * from './useMobile';
 export * from './usePlatform';
+export * from './withPlatform';
 export * from './withMobile';
 export * from './constants';